python/requirements.txt
# ── Core dependencies ─────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9          # Koneksi PostgreSQL
python-dotenv>=1.0.0            # Load .env file

# ── Stemming Bahasa Indonesia ─────────────────────────────────────────────────
PySastrawi>=1.2.0               # Stemmer Bahasa Indonesia (Preprocessing Advanced)

# ── Gemini API (untuk Hybrid Search) ─────────────────────────────────────────
google-generativeai>=0.4.0      # Gemini embedding API

# ── FastAPI (Python API Server) ───────────────────────────────────────────────
fastapi>=0.110.0                # Web framework untuk API
uvicorn>=0.29.0                 # ASGI server untuk FastAPI

# ── Opsional: untuk evaluasi / eksport hasil ─────────────────────────────────
numpy>=1.24.0                   # Operasi vektor numerik (opsional)
pandas>=2.0.0                   # Manipulasi data tabular (evaluasi)
openpyxl>=3.1.0                 # Menulis/menyimpan file Excel (.xlsx)
